♻️ refactor: 테스트 코드 리뷰 반영 #434
- 하드 코딩된 타임 아웃 의미 있는 대기로 변경
- 중복된 주소 검색 및 선택 코드를 fillAddress 함수로 분리
